Move Array.sortBy to helpers file
