+++ panel/src/config/helpers.js
@@ -13,6 +13,27 @@ import sort from "@/helpers/sort.js";
 import string from "@/helpers/string.js";
 import upload from "@/helpers/upload.js";
 
+/**
+ * Array.sortBy()
+ */
+Array.prototype.sortBy = function(sortBy) {
+  const sort = Vue.prototype.$helper.sort();
+  const options = sortBy.split(" ");
+  const field = options[0];
+  const direction = options[1] || "asc";
+
+  return this.sort((a, b) => {
+    const valueA = String(a[field]).toLowerCase();
+    const valueB = String(b[field]).toLowerCase();
+
+    if (direction === "desc") {
+      return sort(valueB, valueA);
+    } else {
+      return sort(valueA, valueB);
+    }
+  });
+};
+
 Vue.prototype.$helper = {
   clone: clone,
   isComponent: isComponent,
